portal local setup script

+++ installer/portal-local-install.sh
@@ -129,13 +129,15 @@ function check_SLI_HOME() {
 }
 function purge_opt() {
    if [ ${PURGE_OPT} == 1 ]; then
-      dialog --title "Deleting ${OPT} directory"  --backtitle "Portal Local Install"  --yesno "Are you sure you want to permanetly delete \"${OPT}\"?" 8 80
-      YES_NO=$?
-      case $YES_NO in
-         0) rm -rf ${OPT}/*;;
-         1) echo "${OPT} directory not deleted";;
-         255) echo "[ESC] key pressed. I am exiting out"; exit;;
-      esac
+      if [ -d ${OPT} ]; then
+         dialog --title "Deleting ${OPT} directory"  --backtitle "Portal Local Install"  --yesno "Are you sure you want to permanetly delete \"${OPT}\"?" 8 80
+         YES_NO=$?
+         case $YES_NO in
+            0) rm -rf ${OPT};;
+            1) echo "${OPT} directory not deleted";;
+            255) echo "[ESC] key pressed. I am exiting out"; exit;;
+         esac
+      fi
    fi
 }
 
@@ -153,7 +155,7 @@ function check_opt() {
 }
 
 function setup_tomcat() {
-   mkdir -p ${PORTAL_TOMCAT}/{webapps,logs,bin} ${DEPLOY_DIR} ${OPT}/staging
+   mkdir -p ${PORTAL_TOMCAT}/{webapps,logs,bin} ${DEPLOY_DIR}
    if [ ! -d ${PORTAL_TOMCAT}/conf ]; then
       NUM=`grep --binary-files=text -n "# END OF SCRIPT #" $0|grep -v cut|cut -d':' -f1`
       NUM=`expr $NUM + 1`
@@ -164,6 +166,11 @@ function setup_tomcat() {
 #   find ${PORTAL_TOMCAT}/webapps -type d -depth 1|grep -v portal |xargs rm -rf
    if [ ! -d ${TOMCAT_HOME} ]; then
       wget -O ${OPT}/apache-tomcat.tar.gz http://apache.mirrors.hoobly.com/tomcat/tomcat-7/v${TOMCAT_VERSION}/bin/apache-tomcat-${TOMCAT_VERSION}.tar.gz
+      RET=$?
+      if [ ${RET} != 0 ]; then
+         echo "Failed to doanload apache-tomcat.  Exiting out..."
+         exit 1
+      fi
       tar -C ${OPT} -zxf ${OPT}/apache-tomcat.tar.gz
       rm -f ${OPT}/apache-tomcat.tar.gz
       if [ -f ${PORTAL_TOMCAT}/conf/server.xml.template ]; then
@@ -194,6 +201,15 @@ chmod a+x ${PORTAL_TOMCAT}/bin/stop.sh
       echo "${TOMCAT_HOME} exists"
       echo "Skipping downloading apache-tomcat-${TOMCAT_VERSION}.tar.gz"
       echo "################################"
+   fi
+   if [ ! -d ${PORTAL_TOMCAT}/webapps/ROOT ]; then
+      cp -fr ${TOMCAT_HOME}/webapps/ROOT ${PORTAL_TOMCAT}/webapps/
+   fi
+   if [ ! -d ${PORTAL_TOMCAT}/webapps/manager ]; then
+      cp -fr ${TOMCAT_HOME}/webapps/manager ${PORTAL_TOMCAT}/webapps/
+   fi
+   if [ ! -d ${PORTAL_TOMCAT}/webapps/host-manager ]; then
+      cp -fr ${TOMCAT_HOME}/webapps/host-manager ${PORTAL_TOMCAT}/webapps/
    fi
 }
 
@@ -260,6 +276,11 @@ app.server.dir=${TOMCAT_HOME}" > ${LIFERAY_HOME}/build.${USER}.properties
    
    if [ ! -f ${PORTAL_TOMCAT}/webapps/portal.war ]; then
       wget -O ${PORTAL_TOMCAT}/webapps/portal.war http://downloads.sourceforge.net/project/lportal/Liferay%20Portal/6.1.0%20GA1/liferay-portal-6.1.0-ce-ga1-20120106155615760.war
+      RET=$?
+      if [ ${RET} != 0 ]; then
+         echo "Failed to doanload Liferay.  Exiting out..."
+         exit 1
+      fi
    else
       echo "################################"
       echo "${PORTAL_TOMCAT}/webapps/portal.war  exists"
@@ -320,9 +341,9 @@ function interactive_app_selection() {
    dialog --backtitle "Portal deployment selection" --radiolist "Select Type of Deployment" 10 40 2 1 "Deploy all" on 2 "Deploy individuals" off 2>/tmp/portal-install.${PID}
    NUM=`sed 's/"//g' /tmp/portal-install.${PID}`
    rm -f /tmp/portal-install.${PID}
-   if [ ${NUM} == 1 ]; then
+   if [ "${NUM}" == "1" ]; then
       deploy_all
-   elif [ ${NUM} == 2 ]; then
+   elif [ "${NUM}" == "2" ]; then
       deploy_individual
    fi
 }
@@ -360,13 +381,8 @@ function update_json() {
          echo '{ "_id" : "213ee853-8983-fe48-bf5e-fde3c3a6437b", "type" : "application", "body" : { "authorized_ed_orgs" : ["IL-SUNSET", "IL", "IL-DAYBREAK", "15", "NC-KRYPTON", "GALACTICA", "CAPRICA", "PICON", "SAGITTARON", "VIRGON", "NY-Parker","NY-Dusk", "NY", "SC-OVERLORD", "KS-GREATVILLE", "KS-SMALLVILLE","KS-OVERLORD"], "version": "0.0", "image_url": "http://placekitten.com/150/150", "administration_url": "http://local.slidev.org:PORTAL_PORT/c/portal/login", "application_url":"http://local.slidev.org:PORTAL_PORT/c/portal/login", "client_secret" : "ghjZfyAXi7qwejklcxziuohiueqjknfdsip9cxzhiu13mnsX", "registration": {"status": "APPROVED", "request_date": 1330521193111, "approval_date": 1330521193111}, "redirect_uri" : "http://local.slidev.org:PORTAL_PORT/portal/c/portal/login", "description" : "Portal Local", "name" : "Portal Local", "is_admin": false, "authorized_for_all_edorgs": true, "allowed_for_all_edorgs": true, "created_by": "slcdeveloper", "installed": false, "client_id" : "lY83c5HmTPX", "behavior": "Full Window App", "vendor" : "SLC"}, "metaData" : { "updated" : { "$date" : 1330521193111 }, "created" : { "$date" : 1330521193111 }} }' |sed "s/PORTAL_PORT/${PORTAL_PORT}/g" >> ${SLI_HOME}/acceptance-tests/test/data/application_fixture.json
          echo '{ "_id" :{ "$binary" : "D0j+8nc6kW62GFLgZnwNnA==", "$type" : "03" }, "type" : "applicationAuthorization", "body" :{ "authId" : "IL-DAYBREAK", "authType" : "EDUCATION_ORGANIZATION", "appIds" : [ "78f71c9a-8e37-0f86-8560-7783379d96f7", "deb9a9d2-771d-40a1-bb9c-7f93b44e51df", "148eebc7-e320-4e35-b7c2-238e90dbd957", "1ad39ff1-65f8-4a16-8912-b49872f1ee96", "63e006bd-fc5f-450f-89c2-69bdfa979c43", "ee3e4e95-0c28-4110-b41b-b29cdec344e6", "25d21fdd-7e97-4aa4-aed0-6d6592a35bb2", "7a8a28cf-f9f2-4ea4-a238-d11b84a3dad2", "8a3def8c-016c-454a-a3ef-cf851e386d4e", "df91814a-ae2e-47f9-b642-742b8c26d65f", "25d21fdd-7e97-4aa4-79d0-6d6592a35bb2", "7a8a28cf-f9f2-4ea4-7938-d11b84a3dad2", "df91814a-ae2e-47f9-7942-742b8c26d65f", "7ae2a2f6-09dc-455d-b139-4d0da6189b52", "af264911-f638-4ec7-b792-69a2a2949b1e", "71ee2a92-788a-4dff-9d72-f1b6f0670aa7", "f4ad653b-2108-4e92-864c-e03722e5ef82", "b26e9c15-54d5-43d5-9d6f-52b8998a6047", "6387e94d-1912-4918-85e0-4f38014253cf", "213ee853-8983-4e40-bf5e-fde3c3a6437b", "52ab3769-1655-4542-971c-1fa02b1b368d", "1ef3990c-e0b5-4046-985c-fbcbab680bb9", "19cca28d-7357-4044-8df9-caad4b1c8ee4", "ee3e4e95-0c28-5114-b41b-b29cdec344e6", "2274ab86-6e21-ab75-9d6f-52b8998a6007", "eab130ba-6e5a-ab75-864c-e03722e523c2", "c09628da-9ea2-d969-85e0-4f3801425388", "c07a656a-6a2c-d9b1-85e0-4fe419425b3b", "1abf3940-84cc-11e1-b0c4-0800200c9a66", "c6251365-e571-482e-8f80-aaece6fd5136", "206e28d3-89a9-db4a-8f80-aaece6fda529", "213ee853-8983-fe48-bf5e-fde3c3a6437b" ] }, "metaData" : { "updated" :{ "$date" : 1332785105123 }, "created" :{ "$date" : 1332785105123 }} }' >> ${SLI_HOME}/acceptance-tests/test/data/applicationAuthorization_fixture.json
       fi
-      echo
-      echo
-      echo
-      echo "######## REQURED ########"
+      echo "######## REQURED if this is your first time installing Portal on your computer ########"
       echo "Please run bundle exec rake realmInitNoPeople"
-      echo
-      echo
       echo
    fi
 }
@@ -395,7 +411,6 @@ function start_tomcat() {
             ${PORTAL_TOMCAT}/bin/stop.sh
             sleep 10
             ${PORTAL_TOMCAT}/bin/start.sh
-            echo "Starting Tomcat again... Please check ${PORTAL_TOMCAT}/logs/catalina.out"
             break
          fi
       done
@@ -484,6 +499,14 @@ fi
 echo "############# BEGIN ###########"
 echo "-Dorg.apache.catalina.loader.WebappClassLoader.ENABLE_CLEAR_REFERENCES=false -Xmx1024m -XX:+CMSClassUnloadingEnabled -XX:+CMSPermGenSweepingEnabled -XX:MaxPermSize=512m -Dsli.encryption.keyStore=${ENCRYPTION_DIR}/ciKeyStore.jks -Dsli.encryption.properties=${ENCRYPTION_DIR}/ciEncryption.properties -Dsli.trust.certificates=${ENCRYPTION_DIR}/trustedCertificates -Dsli.conf=${PORTAL_TOMCAT}/conf/sli.properties -Dliferay.home=${OPT}"
 echo "############# END #############"
+echo "Please check ${PORTAL_TOMCAT}/logs/catalina.out"
+echo "Usefuly hints and commands:"
+echo " Start Tomcat                       : ${PORTAL_TOMCAT}/bin/start.sh"
+echo " Start Tomcat with JPDA (debug mode): ${PORTAL_TOMCAT}/bin/debug-start.sh"
+echo " Stop  Tomcat                       : ${PORTAL_TOMCAT}/bin/stop.sh"
+echo " Location of Tomcat env settings    : ${PORTAL_TOMCAT}/bin/setenv.sh"
+echo " Location of Tomcat log             : ${PORTAL_TOMCAT}/logs/catalina.out"
+echo " Location of Tomcat Web Apss Dir    : ${PORTAL_TOMCAT}/webapps"
 exit
